# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ReminderWorkerResult {
  /** @minimum 0 */
  reminders: number;
  /** @minimum 0 */
  subscriptions: number;
  /** @minimum 0 */
  delivered: number;
  /** @minimum 0 */
  chatDelivered: number;
  /** @minimum 0 */
  failed: number;
  /** @minimum 0 */
  expired: number;
  suppressed: boolean;
}

export interface ReminderWorkerRun {
  ok: boolean;
  runId: string;
  code?: string;
  result?: ReminderWorkerResult;
}

export interface OpenAiImageError {
  status: number;
  message: string;
  /** @nullable */
  code: string | null;
  /** @nullable */
  type: string | null;
  /** @nullable */
  param: string | null;
}

export interface TestImageFailure {
  error: string;
  /** Credential-redacted upstream details, only returned in development */
  openaiError?: OpenAiImageError;
}

export interface TestImageInput {
  /**
     * Image prompt; whitespace-only prompts are not accepted
     * @minLength 1
     * @pattern \S
     */
  prompt: string;
}

export interface AssistantActivity {
  id: number;
  label: string;
  detail: string;
  createdAt: string;
  tone: string;
}

export interface AssistantOverview {
  activeConversationId: number;
  conversationCount: number;
  memoryCount: number;
  connectionCount: number;
  model: string;
  modelStatus: string;
  providerId: string;
  providerStatus: string;
  recentActivity: AssistantActivity[];
}

export interface AssistantProvider {
  id: string;
  name: string;
  vendor: string;
  description: string;
  status: string;
  configured: boolean;
  model: string;
  baseUrlConfigured: boolean;
  credentialConfigured: boolean;
  capabilities: string[];
}

export interface AssistantProviderSelectionInput {
  /** @minLength 1 */
  providerId: string;
}

export interface AssistantProviderSelection {
  activeProviderId: string;
  selectedProviderId: string;
  status: string;
}

export interface AssistantConversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface AssistantConversationInput {
  /** @minLength 1 */
  title: string;
}

export interface AssistantMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
  /** @nullable */
  model: string | null;
  /** @nullable */
  metadata: string | null;
}

export interface AssistantConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: AssistantMessage[];
}

export type ImageCapabilityInputName = typeof ImageCapabilityInputName[keyof typeof ImageCapabilityInputName];


export const ImageCapabilityInputName = {
  generate_image: 'generate_image',
} as const;

export interface ImageCapabilityInput {
  name: ImageCapabilityInputName;
  /**
     * Explicit generate_image capability invocation; not rewritten
     * @minLength 1
     * @pattern \S
     */
  prompt: string;
}

export interface AssistantMessageInput {
  /** Question or message text; may be empty when a photo is attached. */
  content: string;
  /**
     * One resized JPEG, PNG, or WebP photo as a base64 data URL. Never persisted.
     * @maxLength 2800000
     */
  photoDataUrl?: string;
  /** @nullable */
  providerId?: string | null;
  toolCall?: ImageCapabilityInput;
}

export interface AssistantMessagePair {
  userMessage: AssistantMessage;
  assistantMessage: AssistantMessage;
}

export interface AssistantMemory {
  id: number;
  label: string;
  content: string;
  category: string;
  createdAt: string;
}

export interface AssistantMemoryInput {
  /** @minLength 1 */
  label: string;
  /** @minLength 1 */
  content: string;
  category: string;
}

export interface AssistantConnection {
  id: string;
  name: string;
  description: string;
  status: string;
  icon: string;
  available: boolean;
}

export interface HealthStatus {
  status: string;
}