# lib/api-zod/src/generated/types/reminderWorkerResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReminderWorkerResult {
  /** @minimum 0 */
  reminders: number;
  /** @minimum 0 */
  subscriptions: number;
  /** @minimum 0 */
  delivered: number;
  /** @minimum 0 */
  chatDelivered: number;
  /** @minimum 0 */
  failed: number;
  /** @minimum 0 */
  expired: number;
  suppressed: boolean;
}