# lib/api-zod/src/generated/types/reminderWorkerRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReminderWorkerResult } from './reminderWorkerResult';

export interface ReminderWorkerRun {
  ok: boolean;
  runId: string;
  code?: string;
  result?: ReminderWorkerResult;
}